story güncellemesi yapıldı
